requirements.txt
apache-airflow-providers-papermill==3.11.2
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.23.1
azure-keyvault==4.2.0
azure-keyvault-certificates==4.10.0
azure-keyvault-keys==4.11.0
azure-keyvault-secrets==4.10.0
azure-mgmt-core==1.6.0
azure-mgmt-datafactory==9.2.0
azure-storage-blob==12.26.0
azure-storage-file-datalake==12.21.0
duckdb==1.2.2
Flask==2.2.5
imbalanced-learn==0.14.0
imblearn==0.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.7.2
librosa==0.11.0
lxml==5.3.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
numpy==2.2.5
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pendulum==3.1.0
pickleshare==0.7.5
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
ply==3.11
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==21.0.0
pyarrowfs-adlgen2==0.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
regex==2025.7.34
requests==2.32.3
scikit-learn==1.7.1
scipy==1.15.2
selenium==4.32.0
webdriver-manager==4.0.2
Werkzeug==2.3.7
xlrd==2.0.1